requirements.txt
# AutoGen multi-agent framework (0.4+ API)
autogen-agentchat==0.7.5
autogen-ext[openai]==0.7.5

# Azure OpenAI SDK
openai>=2.0.0

# Azure Functions
azure-functions>=1.21.0

# Utilities
python-dotenv>=1.0.1
numpy>=1.26.0

# Document parsing (PDF / DOCX upload support)
pypdf>=4.0.0
python-docx>=1.1.0

# Azure Blob Storage (docs + index in cloud)
azure-storage-blob>=12.21.0
